+++ backend/bin/www
@@ -3,10 +3,10 @@
 /**
  * Module dependencies.
  */
-import app from '../app';
-import debugLib from 'debug';
-import http from 'http';
-const debug = debugLib('your-project-name:server');
+
+var app = require('../app');
+var debug = require('debug')('backend:server');
+var http = require('http');
 
 /**
  * Get port from environment and store in Express.
