+++ backend/package.json
@@ -3,48 +3,12 @@
   "version": "0.0.0",
   "private": true,
   "scripts": {
-    "start": "node ./dist/bin/www",
-    "transpile": "babel ./src --out-dir dist",
-    "clean": "rimraf dist",
-    "build": "npm run clean && npm run transpile",
-    "dev": "NODE_ENV=dev npm run build && npm run start",
-    "prod": "NODE_ENV=prod npm run build && npm run start",
-    "watch": "nodemon",
-    "test": "jest"
+    "start": "node ./bin/www"
   },
   "dependencies": {
-    "@babel/cli": "^7.13.10",
-    "@babel/core": "^7.13.10",
-    "@babel/preset-env": "^7.13.10",
     "cookie-parser": "~1.4.4",
     "debug": "~2.6.9",
     "express": "~4.16.1",
-    "morgan": "~1.9.1",
-    "nodemon": "^2.0.7",
-    "rimraf": "^3.0.2",
-    "sqlite3": "^5.0.2"
-  },
-  "babel": {
-    "presets": [
-      "@babel/preset-env"
-    ]
-  },
-  "nodemonConfig": {
-    "exec": "npm run dev",
-    "watch": [
-      "src/*",
-      "public/*"
-    ],
-    "ignore": [
-      "**/__tests__/**",
-      "*.test.js",
-      "*.spec.js"
-    ]
-  },
-  "devDependencies": {
-    "jest": "^26.6.3"
-  },
-  "jest": {
-    "testEnvironment": "node"
+    "morgan": "~1.9.1"
   }
 }
