+++ backend/routes/index.js
@@ -1,4 +1,4 @@
-import express from 'express';
+var express = require('express');
 var router = express.Router();
 
 /* GET home page. */
@@ -6,4 +6,4 @@ router.get('/', function(req, res, next) {
   res.render('index', { title: 'Express' });
 });
 
-export default router;
+module.exports = router;
