+++ backend/routes/users.js
@@ -1,4 +1,4 @@
-import express from 'express';
+var express = require('express');
 var router = express.Router();
 
 /* GET users listing. */
@@ -6,4 +6,4 @@ router.get('/', function(req, res, next) {
   res.send('respond with a resource');
 });
 
-export default router
+module.exports = router;
